+++ src/main/clojure/cljs/externs.clj
@@ -15,7 +15,7 @@
             CompilerOptions CompilerOptions$Environment SourceFile CompilerInput CommandLineRunner]
            [com.google.javascript.jscomp.parsing Config$JsDocParsing]
            [com.google.javascript.rhino
-            Node Token JSTypeExpression JSDocInfo$Visibility]
+            Node Token JSTypeExpression JSDocInfo JSDocInfo$Visibility]
            [java.util.logging Level]
            [java.net URL]))
 
@@ -61,12 +61,23 @@
   (and (= type :string-lit)
        (= "undefined" value)))
 
+(defn add-prefix
+  "Externs inference uses :prefix meta to both resolve externs as well as generate
+  missing externs information. Google Closure Compiler default externs includes
+  nested types like webCrypto.Crypto. Add prefix information to the returned symbol to
+  simplify resolution later."
+  [type-str]
+  (with-meta (symbol type-str)
+    {:prefix (->> (string/split (name type-str) #"\.")
+               (map symbol) vec)}))
+
 (defn simplify-texpr
   [texpr]
   (case (:type texpr)
-    :string-lit    (some-> (:value texpr) symbol)
-    (:star :qmark) 'any
-    :bang          (simplify-texpr (-> texpr :children first))
+    :string-lit    (-> texpr :value add-prefix)
+    :star          'any
+    ;; TODO: qmark should probably be #{nil T}
+    (:qmark :bang) (simplify-texpr (-> texpr :children first))
     :pipe          (let [[x y] (:children texpr)]
                      (if (undefined? y)
                        (simplify-texpr x)
@@ -77,14 +88,13 @@
   (some-> (.getRoot texpr) parse-texpr simplify-texpr))
 
 (defn params->method-params [xs]
-  (letfn [(not-opt? [x]
-            (not (string/starts-with? (name x) "opt_")))]
-    (let [required (into [] (take-while not-opt? xs))
-          opts (drop-while not-opt? xs)]
-      (loop [ret [required] opts opts]
-        (if-let [opt (first opts)]
-          (recur (conj ret (conj (last ret) opt)) (drop 1 opts))
-          (seq ret))))))
+  (let [not-opt? (complement :optional?)
+        required (into [] (map :name (take-while not-opt? xs)))
+        opts     (map :name (drop-while not-opt? xs))]
+    (loop [ret [required] opts opts]
+      (if-let [opt (first opts)]
+        (recur (conj ret (conj (last ret) opt)) (drop 1 opts))
+        (seq ret)))))
 
 (defn generic? [t]
   (let [s (name t)]
@@ -96,6 +106,18 @@
       (generic? t) 'any
       (= t 'Array) 'array
       :else t)))
+
+(defn get-params
+  "Return param information in JSDoc appearance order. GCL is relatively
+  civilized, so this isn't really a problem."
+  [^JSDocInfo info]
+  (map
+    (fn [n]
+      (let [t (.getParameterType info n)]
+        {:name      (symbol n)
+         :optional? (.isOptionalArg t)
+         :var-args? (.isVarArgs t)}))
+    (.getParameterNames info)))
 
 (defn get-var-info [^Node node]
   (when node
@@ -113,15 +135,15 @@
               (if (or (.hasReturnType info)
                       (as-> (.getParameterCount info) c
                         (and c (pos? c))))
-                (let [arglist  (into [] (map symbol (.getParameterNames info)))
+                (let [arglist  (get-params info)
                       arglists (params->method-params arglist)]
                   {:tag             'Function
                    :js-fn-var       true
                    :ret-tag         (or (some-> (.getReturnType info)
                                           get-tag gtype->cljs-type)
                                         'clj-nil)
-                   :variadic?       (boolean (some '#{var_args} arglist))
-                   :max-fixed-arity (count (take-while #(not= 'var_args %) arglist))
+                   :variadic?       (boolean (some :var-args? arglist))
+                   :max-fixed-arity (count (take-while (complement :var-args?) arglist))
                    :method-params   arglists
                    :arglists        arglists}))))
           {:file *source-file*